+++ src/main/models/index.ts
@@ -1,3 +1,2 @@
 export * from './model-manager';
-export * from './models';
 export * from './types';